Added a new option in AutoSongSegmentLabel. This option can be used to analyze syllables.
Right now, it only shows all the syllable labels and then plots the
individual syllables with the same label in a particular colour. This
makes it possible to visualize the acoustic properties of individual
syllables. It is useful as a way to verify manual labelling too.

The properties of a given syllable as a function of what comes next can
also be visualized.